requirements/bot_requirements.txt
# Bot Framework - Updated compatible version
python-telegram-bot==21.9

# Database - Updated compatible versions
PyMySQL==1.1.0
SQLAlchemy==2.0.35

# AI Integration - Updated stable version
openai==1.51.0

# Environment Variables
python-dotenv==1.0.1

# HTTP Requests
requests==2.31.0

# Date/Time Utilities
pytz==2024.2

# Logging
structlog==24.4.0

# Persian Text Processing
python-bidi==0.4.2
arabic-reshaper==3.0.0

# Type checking
typing-extensions==4.12.2